Add Next/Prior keys, quit on q instead of Escape.
The Next/Prior keys get reported by presenter hardware. Since the Escape
key may be reported (and is very close to the Prior key) I considered it
to dangerous to quit the presentation at a single keypress on Escape.

The keyboard should be hard enough to reach, so we stick with q for now.
Future versions may use a more “complicated” shortcut.